+++ src/aux-common/Files/FileCalculations.ts
@@ -54,6 +54,56 @@ export var ShortId_Length: number = 5;
 export const COMBINE_ACTION_NAME: string = 'onCombine';
 
 /**
+ * The name of the event that represents a file being created.
+ */
+export const CREATE_ACTION_NAME: string = 'onCreate';
+
+/**
+ * The name of the event that represents a file being destroyed.
+ */
+export const DESTROY_ACTION_NAME: string = 'onDestroy';
+
+/**
+ * The name of the event that represents a file being dragged into a context.
+ */
+export const DROP_IN_CONTEXT_ACTION_NAME: string = 'onDropInContext';
+
+/**
+ * The name of the event that represents any file being dragged into a context.
+ */
+export const DROP_ANY_IN_CONTEXT_ACTION_NAME: string = 'onDropAnyInContext';
+
+/**
+ * The name of the event that represents a file being dragged out of a context.
+ */
+export const DRAG_OUT_OF_CONTEXT_ACTION_NAME: string = 'onDragOutOfContext';
+
+/**
+ * The name of the event that represents any file being dragged out of a context.
+ */
+export const DRAG_ANY_OUT_OF_CONTEXT_ACTION_NAME: string = 'onDragAnyOutOfContext';
+
+/**
+ * The name of the event that represents a file being dragged out of the user's inventory.
+ */
+export const DRAG_OUT_OF_INVENTORY_ACTION_NAME: string = 'onDragOutOfInventory';
+
+/**
+ * The name of the event that represents any file being dragged out of the user's inventory.
+ */
+export const DRAG_ANY_OUT_OF_INVENTORY_ACTION_NAME: string = 'onDragAnyOutOfInventory';
+
+/**
+ * The name of the event that represents a file being dropped into the user's inventory.
+ */
+export const DROP_IN_INVENTORY_ACTION_NAME: string = 'onDropInInventory';
+
+/**
+ * The name of the event that represents any file being dropped into the user's inventory.
+ */
+export const DROP_ANY_IN_INVENTORY_ACTION_NAME: string = 'onDropAnyInInventory';
+
+/**
  * Defines an interface for objects that represent assignment formula expressions.
  * Assignment formula expressions are formulas that are only evaluated once.
  * Internally we store them as objects in the tag and display the calculated result.
@@ -665,12 +715,14 @@ export function createFile(id = uuid(), tags: Object['tags'] = {}) {
  * @param id The ID of the new workspace.
  * @param builderContextId The tag that should be used for contexts stored on this workspace.
  */
-export function createWorkspace(id = uuid(), builderContextId: string = createContextId()): Workspace {
+export function createWorkspace(id = uuid(), builderContextId: string = createContextId(), contextType: unknown = true): Workspace {
     
     // checks if given context string is empty or just whitespace
     if(builderContextId.length === 0 || /^\s*$/.test(builderContextId)){
         builderContextId = createContextId();
     }
+
+    
 
     return {
         id: id,
@@ -679,7 +731,7 @@ export function createWorkspace(id = uuid(), builderContextId: string = createCo
             'aux.context.y': 0,
             'aux.context.z': 0,
             [builderContextId]: true,
-            [`${builderContextId}.config`]: '=isBuilder',
+            [`${builderContextId}.config`]: contextType,
             [`${builderContextId}.x`]: 0,
             [`${builderContextId}.y`]: 0,
             [`${builderContextId}.z`]: 0,
@@ -933,7 +985,7 @@ export function getFileInputPlaceholder(calc: FileCalculationContext, file: AuxF
  */
 export function getFileShape(calc: FileCalculationContext, file: File): FileShape {
     const shape: FileShape = calculateFileValue(calc, file, 'aux.shape');
-    if (shape === 'cube' || shape === 'sphere') {
+    if (shape === 'cube' || shape === 'sphere' || shape === 'sprite') {
         return shape;
     }
     return DEFAULT_FILE_SHAPE;