+++ src/aux-server/aux-web/aux-player/interaction/DragOperation/PlayerFileDragOperation.ts
@@ -1,11 +1,12 @@
 import { BaseFileDragOperation } from "../../../shared/interaction/DragOperation/BaseFileDragOperation";
-import { File, FileCalculationContext } from "@casual-simulation/aux-common";
+import { File, FileCalculationContext, DRAG_OUT_OF_INVENTORY_ACTION_NAME, DROP_IN_INVENTORY_ACTION_NAME, FileEvent, DRAG_ANY_OUT_OF_CONTEXT_ACTION_NAME, convertToFormulaObject, DROP_ANY_IN_INVENTORY_ACTION_NAME, DRAG_ANY_OUT_OF_INVENTORY_ACTION_NAME } from "@casual-simulation/aux-common";
 import { PlayerInteractionManager } from "../PlayerInteractionManager";
 import GameView from "../../GameView/GameView";
 import { Intersection, Vector2 } from "three";
 import { Physics } from "../../../shared/scene/Physics";
 import { Input } from "../../../shared/scene/Input";
 import InventoryFile from "../../InventoryFile/InventoryFile";
+import { appManager } from "../../../shared/AppManager";
 
 export class PlayerFileDragOperation extends BaseFileDragOperation {
 
@@ -14,12 +15,15 @@ export class PlayerFileDragOperation extends BaseFileDragOperation {
     // This overrides the base class IGameView
     protected _gameView: GameView;
 
+    private _inInventory: boolean;
+    private _originallyInInventory: boolean;
 
     /**
      * Create a new drag rules.
      */
     constructor(gameView: GameView, interaction: PlayerInteractionManager, files: File[], context: string) {
         super(gameView, interaction, files, context);
+        this._originallyInInventory = this._inInventory = context && appManager.fileManager.userFile.tags._userInventoryContext === context;
     }
 
     protected _onDrag(calc: FileCalculationContext): void {
@@ -33,6 +37,7 @@ export class PlayerFileDragOperation extends BaseFileDragOperation {
                     if (this._context !== vueElement.context) {
                         this._previousContext = this._context;
                         this._context = vueElement.context;
+                        this._inInventory = true;
                     }
 
                     const x = vueElement.slotIndex;
@@ -44,6 +49,7 @@ export class PlayerFileDragOperation extends BaseFileDragOperation {
                 if (this._context !== this._gameView.context) {
                     this._previousContext = this._context;
                     this._context = this._gameView.context;
+                    this._inInventory = false;
                 }
 
                 const mouseDir = Physics.screenPosToRay(this._gameView.getInput().getMouseScreenPos(), this._gameView.getMainCamera());
@@ -61,8 +67,27 @@ export class PlayerFileDragOperation extends BaseFileDragOperation {
                 }
             }
         }
-    }    
+    }
     
-    protected _onDragReleased(): void {
+    protected _onDragReleased(calc: FileCalculationContext): void {
+        super._onDragReleased(calc);
+
+        if (this._originallyInInventory && !this._inInventory) {
+            let events: FileEvent[] = [];
+            let result = appManager.fileManager.helper.actionEvents(DRAG_OUT_OF_INVENTORY_ACTION_NAME, this._files);
+            events.push(...result.events);
+            result = appManager.fileManager.helper.actionEvents(DRAG_ANY_OUT_OF_INVENTORY_ACTION_NAME, null, this._files);
+            events.push(...result.events);
+
+            appManager.fileManager.transaction(...events);
+        } else if (!this._originallyInInventory && this._inInventory) {
+            let events: FileEvent[] = [];
+            let result = appManager.fileManager.helper.actionEvents(DROP_IN_INVENTORY_ACTION_NAME, this._files);
+            events.push(...result.events);
+            result = appManager.fileManager.helper.actionEvents(DROP_ANY_IN_INVENTORY_ACTION_NAME, null, this._files);
+            events.push(...result.events);
+
+            appManager.fileManager.transaction(...events);
+        }
     }
 }