+++ src/aux-server/aux-web/aux-projector/Home/Home.vue
@@ -27,7 +27,6 @@
             <file-table-toggle 
                 :files="selectedFiles" 
                 :raised="true"
-                :showNumFiles="selectionMode !== 'single'"
                 @click="toggleOpen()">
             </file-table-toggle>
           </div>