+++ src/aux-server/aux-web/shared/FileHelper.ts
@@ -98,10 +98,10 @@ export class FileHelper {
     /**
      * Creates a new workspace file.
      */
-    async createWorkspace(builderContextId?: string): Promise<void> {
+    async createWorkspace(builderContextId?: string, contextType?: unknown): Promise<void> {
         console.log('[FileManager] Create File');
 
-        const workspace: Workspace = createWorkspace(undefined, builderContextId);
+        const workspace: Workspace = createWorkspace(undefined, builderContextId, contextType);
         
         await this._tree.addFile(workspace);
     }
@@ -112,19 +112,31 @@ export class FileHelper {
     }
 
     /**
-     * Runs the given event on the given files.
+     * Calculates the list of file events for the given event running on the given files.
      * @param eventName The name of the event to run.
-     * @param files The files that should be searched for handlers for the event name.
+     * @param files The files that should be searched for handlers for the event.
+     * @param arg The argument that should be passed to the event handlers.
      */
-    async action(eventName: string, files: File[]): Promise<void> {
+    actionEvents(eventName: string, files: File[], arg?: any) {
         console.log('[FileManager] Run event:', eventName, 'on files:', files);
 
         // Calculate the events on a single client and then run them in a transaction to make sure the order is right.
-        const fileIds = files.map(f => f.id);
-        const actionData = action(eventName, fileIds, this._userId);
+        const fileIds = files ? files.map(f => f.id) : null;
+        const actionData = action(eventName, fileIds, this._userId, arg);
         const result = calculateActionEvents(this._tree.value, actionData);
         console.log('  result: ', result);
 
+        return result;
+    }
+
+    /**
+     * Runs the given event on the given files.
+     * @param eventName The name of the event to run.
+     * @param files The files that should be searched for handlers for the event name.
+     * @param arg The argument that should be passed to the event handlers.
+     */
+    async action(eventName: string, files: File[], arg?: any): Promise<void> {
+        const result = this.actionEvents(eventName, files, arg);
         await this._tree.addEvents(result.events);
     }