+++ src/aux-server/aux-web/shared/interaction/DragOperation/BaseFileDragOperation.ts
@@ -15,7 +15,13 @@ import {
     getDiffUpdate,
     fileRemoved,
     COMBINE_ACTION_NAME,
-    isMergeable
+    isMergeable,
+    DRAG_OUT_OF_CONTEXT_ACTION_NAME,
+    DROP_IN_CONTEXT_ACTION_NAME,
+    action,
+    calculateActionEvents,
+    DRAG_ANY_OUT_OF_CONTEXT_ACTION_NAME,
+    DROP_ANY_IN_CONTEXT_ACTION_NAME
 } from '@casual-simulation/aux-common';
 
 import { AuxFile3D } from '../../../shared/scene/AuxFile3D';
@@ -39,19 +45,25 @@ export abstract class BaseFileDragOperation implements IOperation {
     protected _other: File;
     protected _context: string;
     protected _previousContext: string;
+    protected _originalContext: string;
+
+    private _inContext: boolean;
 
     /**
      * Create a new drag rules.
-     * @param input the input module to interface with.
-     * @param buttonId the button id of the input that this drag operation is being performed with. If desktop this is the mouse button
+     * @param gameView The game view.
+     * @param interaction The interaction manager.
+     * @param files The files to drag.
+     * @param context The context that the files are currently in.
      */
     constructor(gameView: IGameView, interaction: BaseInteractionManager, files: File[], context: string) {
         this._gameView = gameView;
         this._interaction = interaction;
         this._setFiles(files);
         this._lastScreenPos = this._gameView.getInput().getMouseScreenPos();
-        this._context = context;
+        this._originalContext = this._context = context;
         this._previousContext = null;
+        this._inContext = true;
     }
 
     update(calc: FileCalculationContext): void {
@@ -119,6 +131,7 @@ export abstract class BaseFileDragOperation implements IOperation {
 
     protected _updateFilesPositions(files: File[], gridPosition: Vector2, index: number) {
 
+        this._inContext = true;
         let events: FileEvent[] = [];
         for (let i = 0; i < files.length; i++) {
             let tags = {
@@ -139,6 +152,7 @@ export abstract class BaseFileDragOperation implements IOperation {
     }
 
     protected _updateFileContexts(files: File[], inContext: boolean) {
+        this._inContext = inContext;
         let events: FileEvent[] = [];
         for (let i = 0; i < files.length; i++) {
             let tags = {
@@ -234,11 +248,39 @@ export abstract class BaseFileDragOperation implements IOperation {
         return nextIndex;
     }
 
+    protected _onDragReleased(calc: FileCalculationContext): void {
+        if (this._context !== this._originalContext) {
+
+            let events: FileEvent[] = [];
+            if (this._originalContext) {
+                // trigger drag out of context
+                let result = appManager.fileManager.helper.actionEvents(DRAG_OUT_OF_CONTEXT_ACTION_NAME, this._files, this._originalContext);
+                events.push(...result.events);
+
+                result = appManager.fileManager.helper.actionEvents(DRAG_ANY_OUT_OF_CONTEXT_ACTION_NAME, null, {
+                    context: this._originalContext,
+                    files: this._files
+                });
+                events.push(...result.events);
+            }
+
+            if (this._inContext) {
+                // Trigger drag into context
+                let result = appManager.fileManager.helper.actionEvents(DROP_IN_CONTEXT_ACTION_NAME, this._files, this._context);
+                events.push(...result.events);
+
+                result = appManager.fileManager.helper.actionEvents(DROP_ANY_IN_CONTEXT_ACTION_NAME, null, {
+                    context: this._context,
+                    files: this._files
+                });
+                events.push(...result.events);
+            }
+        }
+    }
 
     //
     // Abstractions
     //
 
     protected abstract _onDrag(calc:FileCalculationContext): void;
-    protected abstract _onDragReleased(calc: FileCalculationContext): void;
 }