+++ src/aux-server/aux-web/shared/scene/decorators/IMeshDecorator.ts
@@ -1,4 +1,4 @@
-import { Mesh, Group } from "three";
+import { Mesh, Group, Sprite } from "three";
 import { AuxFile3D } from "../AuxFile3D";
 import { Event, ArgEvent } from '@casual-simulation/aux-common/Events';
 
@@ -6,7 +6,7 @@ export interface IMeshDecorator {
     
     file3D: AuxFile3D;
     container: Group;
-    mesh: Mesh;
+    mesh: Mesh | Sprite;
 
     /** 
      * Event that gets fired when the mesh is updated.